mdio: Add support c45 modes up to 2018 revision
This adds support for the various speeds and link modes which were
omitted from the original series adding support for c45 registers.

This fixes the 2.5G and 5G PMA types being incorrectly described as
2.25G and 2.5G.

PMA types 0x34 and 0x3d are documented as "BASE-H" and "BASE-T1,"
respectively. This is because they may be used for multiple link modes
(1000BASE-RHA/B/C, 1000BASE-T1, 100BASE-T1, etc.). I have chosen to
leave them as described in the standard, omitting BASE as is customary.

I have left capabilities in the order of their bits. However, it might
be easier to read if they were sorted by speed.

Signed-off-by: Sean Anderson <sean.anderson@seco.com>