Explicit linking behavior
Previously mdio relied on GNU `ld`'s well known secret formula
finding the beginning and end of sections. Compiling with LTO and
optimizations seems to break that.

Now there is a linker script that explains to the linker what is
happening. `externally_visible` was also added to make sure the
cmd functions do not get optimized away since they are not directly
referenced from `main` (might be unnecessary).